requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
attrs
cachetools
certifi
charset-normalizer
discord.py
frozenlist
google-auth
google-auth-httplib2
google-auth-oauthlib
gspread
httplib2
idna
multidict
oauth2client
oauthlib
Pillow
propcache
pyasn1
pyasn1_modules
pyparsing
python-dotenv
pytz
requests
requests-oauthlib
rsa
six
urllib3
yarl
yt-dlp
asyncio
PyNaCl